# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Team Titans - Free Lunch API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/user/bank": {
            "patch": {
                "tags": [
                    "Bank Details"
                ],
                "summary": "Add bank account details",
                "operationId": "211b742d35956aa80b6ad6bd7a38aad8",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "bank_name": {
                                        "type": "string"
                                    },
                                    "bank_number": {
                                        "type": "string"
                                    },
                                    "bank_code": {
                                        "type": "string"
                                    },
                                    "bank_region": {
                                        "type": "string"
                                    },
                                    "currency": {
                                        "type": "string"
                                    },
                                    "currency_code": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "bank_name": "",
                                    "bank_number": "",
                                    "bank_code": "",
                                    "bank_region": "",
                                    "currency": "",
                                    "currency_code": ""
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Add bank account details for the authenticated user",
                                        "value": {
                                            "message": "successfully added bank account details",
                                            "statusCode": 200
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "BAD_REQUEST",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "Add bank account details for the authenticated user",
                                        "value": {
                                            "message": "Error, please try again",
                                            "statusCode": 200
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/bank_details": {
            "get": {
                "tags": [
                    "Bank Details"
                ],
                "summary": "View bank account details",
                "operationId": "62d320a0d6f12aac5e64b802945f3822",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "View bank account details for the authenticated user",
                                        "value": {
                                            "data": [],
                                            "message": "Bank details found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "NOT_FOUND",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "result": {
                                        "summary": "View bank account details for the authenticated user",
                                        "value": {
                                            "message": "No bank details found for the given user."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}